Pull current state into public master for 3.0.12 release.

Highlights:

- GStreamer 1.0 as new multi-media backend.
- libfreenect 0.5 as new PsychKinectCore Kinect backend.
- Ubuntu 14.10 support. Tested with XOrg-1.16, Linux 3.16 and later on radeon, nouveau, intel.
- Improved 10 bpc framebuffer support, new 11 bpc and 16 bpc fb support for AMD gpus.
- Stereo crosstalk reduction support by Diederick.
- Bazillion other new features and fixes i can't remember anymore.
- Improved PsychHID for OSX: Handles multiple parallel kbqueues, maybe copes with MacbookAir 2013+
- MOGL support updated to roughly OpenGL 4.5 (minus unsupportable extensions and omissions).
- Tons of new workarounds for broken Apple operating systems, esp. OSX 10.8/10.9
- Completely drop support for 32-Bit Matlab on all systems, and for OSX 10.7 and earlier.
- Drop official support for Windows Vista and Windows-XP (still works, but no support for problems anymore)
- Prep work for exciting new things to come.

+++ PsychSourceGL/Source/Common/Screen/PsychGLGlue.c
@@ -851,10 +851,10 @@ void PsychDrawDisc(PsychWindowRecordType *windowRecord, float xc, float yc, floa
 	#endif
 
     // Sweep clock-wise over arcAngle degrees, split up into 'numSlices' steps:
-    float step = (arcAngle * 2 * M_PI / 360) / (float) numSlices;
+    float step = (float) ((arcAngle * 2 * M_PI / 360) / (float) numSlices);
 
     // Start at startAngle degrees, where 0 deg. = Upward, 90 deg. = Rightward.
-    startAngle = M_PI/2 - startAngle * 2 * M_PI / 360;
+    startAngle = (float) (M_PI/2 - startAngle * 2 * M_PI / 360);
 
     /* Disable for now, do it slooow ...
        float *vertices = PsychMallocTemp(sizeof(float) * 4 * (numSlices + 1));
