+++ PsychSourceGL/Source/Common/Screen/SCREENFillArc.c
@@ -209,13 +209,13 @@ void PsychRenderArc(unsigned int mode)
     else {
         switch (mode) {
 		case 1: // One pixel thin arc: InnerRadius = OuterRadius - 1
-            PsychDrawDisc(windowRecord, (float) cx, (float) cy, (w/2) - 1.0, w/2, (int) w, (float) 1, (float) -h/w, (float) *startAngle, (float) *arcAngle);
+            PsychDrawDisc(windowRecord, (float) cx, (float) cy, (float) ((w/2) - 1.0), (float) w/2, (int) w, (float) 1, (float) -h / (float) w, (float) *startAngle, (float) *arcAngle);
 			break;
 		case 2: // dotSize thick arc:  InnerRadius = OuterRadius - dotsize
-            PsychDrawDisc(windowRecord, (float) cx, (float) cy, (dotSize < (w/2)) ? (w/2) - dotSize : 0, w/2, (int) w, (float) 1, (float) -h/w, (float) *startAngle, (float) *arcAngle);
+            PsychDrawDisc(windowRecord, (float) cx, (float) cy, (float) ((dotSize < (w/2)) ? (w/2) - dotSize : 0), (float) w / 2, (int) w, (float) 1, (float) -h/ (float) w, (float) *startAngle, (float) *arcAngle);
 			break;
 		case 3: // Filled arc:
-            PsychDrawDisc(windowRecord, (float) cx, (float) cy, 0, w/2, (int) w, (float) 1, (float) -h/w, (float) *startAngle, (float) *arcAngle);
+            PsychDrawDisc(windowRecord, (float) cx, (float) cy, (float) 0, (float) w / 2, (int) w, (float) 1, (float) -h / (float) w, (float) *startAngle, (float) *arcAngle);
 			break;
         }
     }
